Fix crash on multiple threads.
The extension crashes when used in multiple threads. The problem is in UDP_CheckProc when it is determining whether it should read the socket. If there are multiple threads, the order in which the threads call this function is not deterministic, so it is possible that the wrong thread will read the socket.

Bug fix from http://fossil.etoyoc.com/fossil/tcludp/tktview?name=b3905a8226