Fix for crash closing a udp channel on Windows.
The problem is encountered when:

- a readable fileevent is setup on the channel
- the channel is closed
- the program calls vwait to go back into the event loop and waits for some flag to be set (from a different event source).

Then a queued file readable callback for the deleted channel can be
executed from within the vwait, causing the crash. Maybe the crash
could also occur using the normal wish event loop, but I have only
seen it with a vwait following the channel close.

The patch uses Tcl_DeleteEvents to remove queued events for a channel when it is closed.

Patch fix from https://sourceforge.net/p/tcludp/patches/14/